Fix critical bugs in trait extraction
1. Add missing {question} and {answer} placeholders to 13 trait eval_prompts
   - Judge was scoring without seeing actual responses
   - Added placeholders to end of eval_prompt in all affected traits

2. Fix dtype mismatch in GradientMethod
   - vector initialization now matches activation dtype (float16)
   - Prevents 'expected Half but found Float' errors

These bugs prevented proper judging and gradient extraction.

+++ traitlens/methods.py
@@ -278,7 +278,7 @@ class GradientMethod(ExtractionMethod):
         hidden_dim = pos_acts.shape[1]
 
         # Initialize vector randomly
-        vector = torch.randn(hidden_dim, device=pos_acts.device, requires_grad=True)
+        vector = torch.randn(hidden_dim, device=pos_acts.device, dtype=pos_acts.dtype, requires_grad=True)
 
         optimizer = torch.optim.Adam([vector], lr=lr)
         loss_history = []
